refactor: split TaskInfoPanel into focused modules
